+++ Examples/ObjectTypes/Example_ObjectTypeBuilder.ps1
@@ -1,6 +1,8 @@
 Test-Module PSInsight
 
-$SchemaID = $(Get-InsightObjectSchema | Where { $_.name -eq 'CompanyName'}).id
-$JSONpath = '.\PSInsight\Examples'
-$Objects = Get-Content "$JSONpath\ObjectTypes.json" -Raw | ConvertFrom-Json
-Build-ObjectType -ObjectArray $Objects -SchemaID $SchemaID -InsightApiKey ''
+$SchemaName = 'CompanyName'
+$Global:InsightApiKey = Get-Secret -Name 'Insight_API' -AsPlainText
+$SchemaID = $(Get-InsightObjectSchema -InsightApiKey $InsightApiKey | Where { $_.name -eq $SchemaName}).id
+$JSONpath = '.\PSInsight\Examples\ObjectTypes'
+$Objects = Get-Content -Path "$JSONpath\ObjectTypes.json" -Raw | ConvertFrom-Json
+Build-ObjectType -ObjectArray $Objects -SchemaID $SchemaID -InsightApiKey $InsightApiKey